FIX: Building grammers properly

+++ build.py
@@ -10,10 +10,10 @@ from shutil import move, rmtree
 from contextlib import chdir
 from distutils.dir_util import copy_tree
 
-HELIX_VERSION = "23.03"
+HELIX_VERSION = "23.05"
 DEBIAN_DIRECTORY = os.path.join(os.path.dirname(os.path.abspath(__file__)), "debian")
 TARGET_DIRECTORY = os.path.join(os.getcwd(), "target")
-HELIX_SOURCE_CODE_URL = "https://github.com/helix-editor/helix/archive/refs/tags"
+HELIX_SOURCE_CODE_URL = f"https://github.com/helix-editor/helix/releases/download/{HELIX_VERSION}/helix-{HELIX_VERSION}-source.tar.xz"
 
 
 def get_args() -> tuple[str, str]:
@@ -35,20 +35,17 @@ Example:
 
 
 def prepare_target(target_directory: str) -> None:
-    rmtree(target_directory)
+    if os.path.exists(target_directory):
+        rmtree(target_directory)
+
     os.mkdir(target_directory)
 
 
-def get_helix_source_url(helix_version: str) -> str:
-    return f"{HELIX_SOURCE_CODE_URL}/{helix_version}.tar.gz"
-
-
 def download_helix_release(target_directory_path: str, helix_version: str) -> str:
-    url = get_helix_source_url(helix_version)
-    filename = f"helix_{helix_version}.orig.tar.gz"
+    filename = f"helix_{helix_version}.orig.tar.xz"
 
-    print(f"-> Downloading {url}...")
-    response = requests.get(url, allow_redirects=True)
+    print(f"-> Downloading {HELIX_SOURCE_CODE_URL}...")
+    response = requests.get(HELIX_SOURCE_CODE_URL, allow_redirects=True)
 
     target_file_path = os.path.join(target_directory_path, filename)
 
@@ -86,29 +83,10 @@ def create_cargo_vendor_archive(
         subprocess.check_call(["tar", "cJf", tar_file_path, "vendor/"])
 
 
-def crate_grammar_sources_archive(
-    source_directory_path: str, destination_directory_path: str
-) -> None:
-    with chdir(source_directory_path):
-        print("-> Run cargo build...")
-        subprocess.check_call(["cargo", "build", "--release"])
-
-        tar_file_path = os.path.join(destination_directory_path, "grammars.src.tar.xz")
-        print(f"-> Create {tar_file_path}...")
-        subprocess.check_call(
-            ["tar", "cJf", tar_file_path, "runtime/grammars/sources/"]
-        )
-
-
 def create_dependencies_archives(
     source_directory_path: str, destination_directory_path: str
 ) -> None:
     create_cargo_vendor_archive(
-        source_directory_path=source_directory_path,
-        destination_directory_path=destination_directory_path,
-    )
-
-    crate_grammar_sources_archive(
         source_directory_path=source_directory_path,
         destination_directory_path=destination_directory_path,
     )
@@ -118,6 +96,7 @@ def prepare_for_build(
     source_directory_path: str
 ) -> str:
     prepare_target(TARGET_DIRECTORY)
+    prepare_target(source_directory_path)
 
     release_file_path = download_helix_release(
         target_directory_path=TARGET_DIRECTORY,
@@ -125,7 +104,7 @@ def prepare_for_build(
     )
 
     unarchive_helix_release(
-        target_directory_path=TARGET_DIRECTORY,
+        target_directory_path=source_directory_path,
         archive_file_path=release_file_path,
     )
 
@@ -164,7 +143,7 @@ def update_changelog(source_directory_path, ubuntu_codename, changelog_version) 
 def run_debuild(source_directory_path: str) -> None:
     with chdir(source_directory_path):
         print("-> Run debuild...")
-        subprocess.check_call(["debuild", "-S", "-sa"])
+        subprocess.check_call(["debuild", "--no-lintian", "-S", "-sa"])
 
 
 def run_build(
@@ -179,7 +158,7 @@ def run_build(
     prepare_target(source_directory_path)
 
     unarchive_helix_release(
-        target_directory_path=TARGET_DIRECTORY, archive_file_path=release_file_path
+        target_directory_path=source_directory_path, archive_file_path=release_file_path
     )
 
     move_debian_files(
